Fix layout issues in CoreConfigWizard; cleanup
Qt has trouble layouting widgets dynamically added to wizard pages,
so switch to using QStackedWidget for storage and authenticator
settings. Add size policies as needed and explicitly set the size of
all pages to fit the largest one to ensure that the wizard is large
enough.

While we're at it, clean up related code a bit.

Closes GH-170.